refactor: share helper fix
